+++ ChangeLog.md
@@ -26,6 +26,8 @@
 - The file names of client certificates created within Rogallo now use a
   uuid4 for the name, rather than the host/port of the initially-associated
   site. ([#373](https://github.com/davep/rogallo/pull/373))
+- Consider command inputs in `user@example.com` form to be `finger`
+  requests. ([#381](https://github.com/davep/rogallo/pull/381))
 
 ## v1.12.1